src/requirements.txt
docker
discord-webhook
pyyaml
schedule